Created flask_notworking and moved files there that doesn't work currently
